api/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
async-property==0.2.2
cached-property==1.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
deprecation==2.1.0
ecdsa==0.19.1
fastapi==0.115.11
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
jwcrypto==1.5.6
packaging==24.2
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.4.0
python-keycloak==5.3.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
